added risingwave median udf

+++ container/risingwave_udf_median/median.py
@@ -7,7 +7,7 @@ import socket
 
 
 # Define a table function
-@udf(input_types=['numeric'], result_type='numeric')
+@udf(input_types=['double'], result_type='double')
 def median(n):
     return statistics.median(n)
